restructure manuscript section file tree
