+++ src/components/organisms/SidePanel/ManuscriptSidePanel/Tabs/History/history.scss
